factored wallets into separate view
